requirements.txt
celery==4.4.2
Pillow==7.0.0
flower==0.9.3
braintree==3.59.0
WeasyPrint==51
markdown==3.2.1
django-rosetta==0.9.3
django-parler==2.0.1
zeep
pip install pycairo
